+++ Kernel/System/CustomerUser/DB.pm
@@ -1264,10 +1264,12 @@ sub CustomerUserDataGet {
         $CustomerUserListFieldsMap = [ 'first_name', 'last_name', 'email', ];
     }
 
-    # to build the UserMailString
-    my %LookupCustomerUserListFields = map { $_ => 1 } @{$CustomerUserListFieldsMap};
-    my @CustomerUserListFields
-        = map { $_->[0] } grep { $LookupCustomerUserListFields{ $_->[2] } } @{ $Self->{CustomerUserMap}->{Map} };
+    # Order fields by CustomerUserListFields (see bug#13821).
+    my @CustomerUserListFields;
+    for my $Field ( @{$CustomerUserListFieldsMap} ) {
+        my @FieldNames = map { $_->[0] } grep { $_->[2] eq $Field } @{ $Self->{CustomerUserMap}->{Map} };
+        push @CustomerUserListFields, $FieldNames[0];
+    }
 
     my $UserMailString = '';
     my @UserMailStringParts;